requirements.txt.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.2
scikit-learn>=1.0.2
Pillow>=9.0.0
